fix : ~Count타입을 int->long 변경

+++ src/main/java/com/redhorse/deokhugam/domain/review/mapper/ReviewMapper.java
@@ -23,5 +23,5 @@ public interface ReviewMapper {
   @Mapping(source = "review.user.nickname", target = "userNickname")
   @Mapping(source = "likedByMe", target = "likedByMe")
   @Mapping(source="commentCount",target="commentCount")
-  ReviewDto toDto(Review review, boolean likedByMe, int commentCount);
+  ReviewDto toDto(Review review, boolean likedByMe, long commentCount);
 }
